CASH-922
* Separate watches on active loan only so that it doesn’t run on every card instance.
* Fix js errors when navigating to a category page then back to lend-by-category
* Cleanup unused variables
* Guard code
